+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Wherever possible enums are used for values
 * Where UUIDs are expected the types in the API have changed from String to UUID.  You can wrap strings in `UUID.fromString()` if necessary.
 * `prepareUpload()` methods have been moved to a dedicated `PrepareUploadHelper` class and now use a regular Map to store parameters 
+* Error messages from the API are now accessible via `NotificationClientHttpException`
 
 ## 5.2.0-RELEASE
 * Added fields related to cost data in response: